requirements.txt
accelerate==0.27.2
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
datasets==2.16.1
dill==0.3.7
docker-pycreds==0.4.0
einops==0.7.0
evaluate==0.4.1
filelock==3.13.1
fonttools==4.51.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
huggingface-hub==0.20.1
hydra-core==1.3.2
idna==3.6
install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lightning==2.1.3
lightning-utilities==0.10.0
MarkupSafe==2.1.3
matplotlib==3.9.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
numpy==1.26.2
omegaconf==2.3.0
packaging==23.2
pandas==2.1.4
pillow==10.3.0
protobuf==4.25.2
psutil==5.9.7
pyarrow==14.0.2
pyarrow-hotfix==0.6
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.8.2
pytorch-lightning==2.1.3
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
responses==0.18.0
safetensors==0.4.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==1.40.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.19.1
torch==2.1.2
torchmetrics==1.3.0
tqdm==4.66.1
transformers==4.40.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wandb==0.16.2
xxhash==3.4.1
yarl==1.9.4
